TypeBox 0.29.0 Not Type (#22)
## Summary

Hey! This PR goes ahead and implements the TypeBox 0.29.0 Not type (as
discussed on
https://github.com/xddq/schema2typebox/pull/21#issuecomment-1615823630).
This update as been fast forwarded on the TypeBox repository along with
some other bug fixes related to TypeScript 5.1.6 / Indexed Types (so
seemed good to roll the change through now). Submitting this PR to patch
`schema2typebox` to the latest version.

The change log notes for the 0.29.0 release can be found
https://github.com/sinclairzx81/typebox/blob/master/changelog/0.29.0.md

Cheers!